% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stargazer.R
\name{capture_stargazer_tab}
\alias{capture_stargazer_tab}
\title{Capture everything within the tabular environment of stargazer output}
\usage{
capture_stargazer_tab(star_tbl)
}
\arguments{
\item{star_tbl}{Stargazer object}
}
\value{
Character vector
}
\description{
\code{capture_stargazer_tab} is a function that extracts everything inside
the tabular environment of stargazer output.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
