% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{is_tex}
\alias{is_tex}
\title{Check if file path has tex extension}
\usage{
is_tex(path)
}
\arguments{
\item{path}{Character vector with path}
}
\value{
TRUE if extension is tex and FALSE if not
}
\description{
\code{is_tex} is a function that checks if a file path has a tex extension
}
\details{
Used in scripts: {}
Robust function: FALSE
}
