% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{ltx_caption}
\alias{ltx_caption}
\title{Customise latex output}
\usage{
ltx_caption(x, tbl_note = "", print_tbl = FALSE)
}
\arguments{
\item{x}{Output from \code{texreg::texreg}, \code{kableExtra::kable},
\code{stargazer::stargazer} etc.}

\item{tbl_note}{Character string with explanatory text for the table. This
is placed under the table title.}

\item{print_tbl}{TRUE/FALSE. Should the table be printed to console?}
}
\value{
This function prints out the modified latex code
}
\description{
\code{ltx_caption} is a function that modifies latex output from
packages such as texreg, kable and stargazer. What it does is that it removes
the usepackage commands added to the output if options such as booktabs =
TRUE is used (in texreg package). Furthermore, it changes the format of the
title and adds the option of including a table text directly below the main
title of the table.
Used in scripts: {}
Robust function: TRUE
}
