% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{ltx_decimals}
\alias{ltx_decimals}
\title{Remove decimals from latex code}
\usage{
ltx_decimals(x, n_digits = 3, print_tbl = FALSE)
}
\arguments{
\item{x}{Latex code either in form of string or output from other functions}

\item{n_digits}{When to remove decimals, e.g., n_digits = 3 means for numbers
greater than 99.}

\item{print_tbl}{TRUE/FALSE. Should the table be printed to console?}
}
\value{
Returns the input \code{x} with decimals removed. It prints out the
output to console as well as return it invisibly for easy use in pipes.
}
\description{
\code{ltx_decimals} is a function that removes decimals for numbers
in latex code that are bigger than n_digits. For instance if n_digits = 2,
and in the latex code there is a number, 23.34, this function will turn it
into 23.
Used in scripts: {}
Robust function: TRUE
}
