% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_environment.R
\name{ltx_environment}
\alias{ltx_environment}
\title{Deprecated: ltx_environment function}
\usage{
ltx_environment(x, print_tbl = FALSE)
}
\arguments{
\item{x}{A character vector representing LaTeX code for a table.}

\item{print_tbl}{Logical, whether to print the modified LaTeX table
to the console (default is \code{FALSE}).}
}
\value{
Returns the modified LaTeX code as a character vector (invisible).
}
\description{
This function is deprecated and should not be used anymore.
It remains available for backward compatibility.
}
\details{
This function modifies the LaTeX code by changing the table environment from
\code{tabular} to \code{tabularx}. However, it is deprecated and should no longer be
used.
}
\seealso{
\code{\link[=kable_wrapper]{kable_wrapper()}}
}
