% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{ltx_multipanel}
\alias{ltx_multipanel}
\title{Create two-panel regression table}
\usage{
ltx_multipanel(x, print_tbl = FALSE)
}
\arguments{
\item{x}{List of regression table latex code (output from texreg)}

\item{print_tbl}{TRUE/FALSE. Should latex code be printed to console?}
}
\value{
Latex code for multipanel table
}
\description{
\code{ltx_multipanel} is a function that takes a list \code{x} of two
latex tbls and merges them together such that they form a two-panel table.
Used in scripts: {}
Robust function: FALSE
}
