% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{ltx_placement}
\alias{ltx_placement}
\title{Set table placement for Latex table}
\usage{
ltx_placement(x, tbl_placement = NULL, print_tbl = FALSE)
}
\arguments{
\item{x}{Character string with latex code}

\item{tbl_placement}{Character specifying table placement. Must conform to
standard options in latex}

\item{print_tbl}{TRUE/FALSE. Should the table be printed to console?}
}
\value{
Character string with latex code. Prints it out to console and
returns the text in strings to use with pipes.
}
\description{
\code{ltx_placement} lets the user determine the placement of a latex
table. The options from kableExtra is limited and this gives more control.
The option of hold_position in \code{kable_styling} produces a warning in
Overleaf that the requirement is to strong.
Used in scripts: {}
Robust function: TRUE
}
