% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_print_tbl.R
\name{ltx_print_tbl}
\alias{ltx_print_tbl}
\title{Print and Copy LaTeX Table to Clipboard}
\usage{
ltx_print_tbl(x, print = FALSE)
}
\arguments{
\item{x}{A character string representing LaTeX table or text that you wish
to print and copy to the clipboard.}

\item{print}{Logical. If \code{TRUE}, the LaTeX table or text in \code{x} will be
printed to the console. Defaults to \code{FALSE}.}
}
\value{
The function does not return any value. It performs the side effects of
printing to the console (if \code{print} is \code{TRUE}) and copying \code{x} to the
clipboard.
}
\description{
The \code{ltx_print_tbl} function prints a LaTeX table or text to the console and
copies the content to the system clipboard. This is useful for quickly
copying and pasting LaTeX code into a document while also optionally printing
it to the console.
}
\details{
The function is designed to streamline the process of working with LaTeX
tables or other formatted text. When \code{print} is set to \code{TRUE}, the function
prints the content of \code{x} to the console with line breaks around the table
or text for visual clarity. It also copies the content to the system
clipboard using \code{utils::writeClipboard()}, making it easy to paste directly
into LaTeX documents.
}
\examples{
\dontrun{
# Example LaTeX table string
tbl <- "\\\\begin{tabular}{|c|c|}\n\\\\hline\nA & B \\\\\\\\\n\\\\hline\n1 & 2 \\\\\\\\\n\\\\hline\n\\\\end{tabular}"

# Print and copy the table to clipboard
ltx_print_tbl(tbl, print = TRUE)
}
}
\seealso{
\code{\link[utils:clipboard]{utils::writeClipboard()}}
}
