% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{ltx_resize}
\alias{ltx_resize}
\title{Add resizebox to latex table}
\usage{
ltx_resize(x, print_tbl = FALSE)
}
\arguments{
\item{x}{Character string of latex code}

\item{print_tbl}{TRUE/FALSE. Should the table be printed to console?}
}
\value{
Returns a character string of latex code.
}
\description{
\code{ltx_resize} is a function that adds the resizebox command to
latex table code in order to size tables to fit a page. This function is
mostly intended for use with output from regression tables (texreg package)
as this does not have an option to resize the table using resizebox.
Used in scripts: {}
Robust function: TRUE
}
