% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_round_numbers.R
\name{ltx_round_numbers}
\alias{ltx_round_numbers}
\title{Round Numbers in LaTeX Tables}
\usage{
ltx_round_numbers(
  latex_table,
  threshold = 10,
  big_dec = 0,
  small_dec = 2,
  print_tbl = FALSE
)
}
\arguments{
\item{latex_table}{A character vector, data frame, or matrix representing
a LaTeX table. The function will round numerical values within the LaTeX
code.}

\item{threshold}{A numeric value that determines the cutoff for rounding.
Numbers greater than or equal to this value will be rounded to \code{big_dec}
decimal places, while those below will be rounded to \code{small_dec} decimal
places.}

\item{big_dec}{An integer indicating the number of decimal places to round
to for numbers greater than or equal to the threshold. Default is 0.}

\item{small_dec}{An integer indicating the number of decimal places to round
to for numbers below the threshold. Default is 2.}

\item{print_tbl}{A logical flag indicating whether the modified table should
be printed to the console and copied to the clipboard. Defaults to \code{FALSE}.}
}
\value{
A character vector representing the modified LaTeX table with rounded
numbers.
}
\description{
This function takes a LaTeX table with numerical data and rounds numbers
based on a specified threshold. Numbers greater than or equal to the threshold
will be rounded to a specified number of decimal places (big_dec), while
numbers below the threshold will be rounded to a different number of decimal
places (small_dec).
}
\examples{
\dontrun{
latex_table <- c(
  "\\\\begin{tabular}{|c|c|c|}",
  "\\\\hline",
  "Value A & Value B & Value C \\\\\\\\",
  "\\\\hline",
  "1.234 & 10.5678 & 12345.6789 \\\\\\\\",
  "2.345 & 9.8765 & 6789.0123 \\\\\\\\",
  "\\\\hline",
  "\\\\end{tabular}"
)

rounded_table <- ltx_round_numbers(latex_table, threshold = 10)
print(rounded_table)
}
}
