% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_stretch.R
\name{ltx_stretch}
\alias{ltx_stretch}
\title{Apply Array Stretch to a LaTeX Table}
\usage{
ltx_stretch(x, array_stretch = 1.2, verbose = TRUE, print_tbl = FALSE)
}
\arguments{
\item{x}{A character vector representing a LaTeX table or an object that can
be captured as LaTeX output using the \code{ltx_capture_output()} function.}

\item{array_stretch}{Numeric. The row stretch factor to apply.
Must be a positive number. Default is \code{1.2}.}

\item{verbose}{Logical. If \code{TRUE}, prints a message indicating the applied
array stretch. Default is \code{TRUE}.}

\item{print_tbl}{Logical. If \code{TRUE}, prints the modified LaTeX table to the
console. Default is \code{FALSE}.}
}
\value{
Returns the modified LaTeX table as a character vector with the
array stretch applied. Invisibly returns the modified table.
}
\description{
This function modifies a LaTeX table to adjust the row stretch
(i.e., vertical space between rows) by appending the array stretch command
(\verb{\\ra\{\}}) after the beginning of a \code{tabular} or \code{tabularx} environment.
The \verb{\\ra\{\}} command is custom defined in my Overleaf as
\verb{\newcommand{\ra}[1]\{\renewcommand{\arraystretch}{#1}\}}.
}
\details{
The function searches for \verb{\\begin\{table\}} in the LaTeX table and
appends the array stretch command \verb{\\ra\{\}} right after. If no such environment
is found, an error is thrown. The function does not modify the
\verb{\\begin\{table\}} environment.
}
\examples{
\dontrun{
# Example LaTeX table
latex_table <- c("\\\\begin{table}",
"\\\\begin{tabular}{ccc}",
"\\\\hline", "1 & 2 & 3 \\\\\\\\",
"\\\\hline",
"\\\\end{tabular}",
"\\\\end{table}")
# Apply array stretch of 1.5
modified_table <- ltx_stretch(latex_table, array_stretch = 1.5, verbose = TRUE)

# Print the modified table
ltx_stretch(latex_table, array_stretch = 1.5, print_tbl = TRUE)
}
}
