% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_tidy_table.R
\name{ltx_tidy_table}
\alias{ltx_tidy_table}
\title{Tidy and Format LaTeX Table}
\usage{
ltx_tidy_table(
  x,
  tbl_note = "",
  replace_idx = 1,
  array_stretch = 1.2,
  threshold = 10,
  big_dec = 0,
  small_dec = 2,
  print_tbl = TRUE
)
}
\arguments{
\item{x}{A character vector or list containing the LaTeX table input.}

\item{tbl_note}{A string representing the caption to be added to the table.
Defaults to "".}

\item{replace_idx}{A numeric value representing the index at which to replace
the table placement. Defaults to 1.}

\item{array_stretch}{A numeric value to set the \verb{\\\\arraystretch} for row
spacing. Defaults to 1.2.}

\item{threshold}{A numeric value defining the threshold for rounding large
numbers. Numbers greater than or equal to this will use \code{big_dec}.
Defaults to 10.}

\item{big_dec}{Number of decimal places for large numbers (>= threshold).
Defaults to 0.}

\item{small_dec}{Number of decimal places for small numbers (< threshold).
Defaults to 2.}

\item{print_tbl}{Logical, if \code{TRUE}, the modified LaTeX table will be printed
to the console. Defaults to \code{TRUE}.}
}
\value{
The processed LaTeX table is returned invisibly.
If \code{print_tbl = TRUE}, the table is printed to the console.
}
\description{
This function processes a LaTeX table by applying a series of transformations:
adding a caption, adjusting placement, setting row spacing, rounding numbers,
and wrapping numbers in math mode. It streamlines the formatting of LaTeX tables
to ensure consistency in presentation and compliance with LaTeX math mode formatting.
}
\details{
The function works by applying the following steps:
\itemize{
\item \code{ltx_capture_output(x)}: Converts the input LaTeX table into a manipulable format.
\item \code{ltx_caption(tbl_note)}: Adds a caption to the LaTeX table.
\item \code{ltx_placement(replace_idx)}: Adjusts the LaTeX table placement.
\item \code{ltx_stretch(array_stretch)}: Sets the row spacing for the table by modifying the \verb{\\\\arraystretch}.
\item \code{ltx_round_numbers(threshold, big_dec, small_dec)}: Rounds the numbers in the table based on the given threshold.
\item \code{ltx_wrap_table_math()}: Wraps the numbers in LaTeX math mode (\verb{$number$}) where appropriate.
}
}
\examples{
\dontrun{
# Example LaTeX input
latex_input <- c("\\\\begin{table}", "\\\\begin{tabular}{|c|c|}", "\\\\hline",
                 "Value & 1234 \\\\\\\\", "Temperature & 98.6 \\\\\\\\", "\\\\hline",
                 "\\\\end{tabular}", "\\\\end{table}")

# Tidy the table and print it
ltx_tidy_table(latex_input, tbl_note = "Sample Table", array_stretch = 1.5)
}

}
