% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tex.R
\name{write_tex}
\alias{write_tex}
\title{Write a TeX File}
\usage{
write_tex(x, path, verbose = TRUE)
}
\arguments{
\item{x}{A character vector to write to the file.}

\item{path}{A character string specifying the file path. Should be a
valid path for a .tex file.}

\item{verbose}{A logical value indicating whether to print messages and
warnings. Default is TRUE.}
}
\value{
The function returns the input character vector invisibly.
}
\description{
This function writes a character vector to a .tex file. It ensures the
specified path has a .tex extension, checks for directory existence, and
allows for overwriting existing files with user feedback.
}
\examples{
\dontrun{
# Write a simple LaTeX document
write_tex(c("\\\\documentclass{article}", "\\\\begin{document}", "Hello, world!", "\\\\end{document}"),
           "output/document.tex")
}
}
