% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tex.R
\name{is_tex}
\alias{is_tex}
\title{Check if the File is a TeX File}
\usage{
is_tex(path)
}
\arguments{
\item{path}{A character string specifying the file path.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE}) indicating whether the file has
a \code{.tex} extension.
}
\description{
This function checks whether the specified file path has a \code{.tex} file
extension.
}
\examples{
\dontrun{
is_tex("document.tex")  # Returns TRUE
is_tex("image.png")     # Returns FALSE
}
}
