% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{ltx_environment}
\alias{ltx_environment}
\title{Change default table environment for latex tables}
\usage{
ltx_environment(x, print_tbl = FALSE)
}
\arguments{
\item{x}{Character vector (ouput from stargazer/kable) - latex table code}

\item{print_tbl}{TRUE/FALSE. Should the table be printed to console?}
}
\value{
Returns latex code with a new table environment
}
\description{
\code{ltx_environment} is a function designed to take output from
stargazer or kable and change the table environment from \emph{tabular} to
\emph{tabularx}, and also specify that the table should fill the entire textwidth.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
