% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{ltx_fontsize}
\alias{ltx_fontsize}
\title{Set table fontsize}
\usage{
ltx_fontsize(x, font_size = 8, print_tbl = FALSE)
}
\arguments{
\item{x}{Character string of latex code}

\item{font_size}{Numeric, fontsize in latex unit \emph{pt}}

\item{print_tbl}{TRUE/FALSE. Should the table be printed to console?}
}
\value{
Returns a character string of latex code.
}
\description{
\code{ltx_fontsize} adds the latex command fontsize{}{} to the latex
table allowing one to adjust the fontsize.
Used in scripts: {}
Robust function: TRUE
}
