% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx_resize.R
\name{ltx_resize}
\alias{ltx_resize}
\title{Resize LaTeX Tabular Environments}
\usage{
ltx_resize(x, print_tbl = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{A character vector representing the LaTeX code that contains tabular
environments.}

\item{print_tbl}{A logical flag indicating whether the modified table should
be printed to the console and copied to the clipboard. Defaults to \code{FALSE}.}

\item{verbose}{A logical flag to indicate whether messages should be printed
to the console. Defaults to \code{TRUE}.}
}
\value{
The modified LaTeX code with the \code{resizebox} applied around tabular
environments, returned as a character vector. If \code{print_tbl} is \code{TRUE},
the modified table is printed to the console.
}
\description{
This function wraps LaTeX tabular-like environments
(e.g., \code{tabular}, \code{tabularx}, \code{longtable}) with a \verb{\\resizebox} command to
resize them to fit the \verb{\\linewidth}. The function excludes resizing for
\code{table} environments to prevent undesired behavior. This is useful for
ensuring tables are properly scaled within the document margins in LaTeX.
}
\details{
The function works by identifying lines containing \verb{\\begin\{tabular\}},
\verb{\\begin\{tabularx\}}, or other supported tabular environments
(excluding \code{table}), and wraps these environments with a
\verb{\\resizebox\{\\\\linewidth\}\{!\}\{...\}} command. It ensures that the
\verb{\\end\{tabular\}} line is closed properly.
}
\examples{
\dontrun{
# Example LaTeX table code
x <- c("\\\\begin{tabular}{ll}", "A & B \\\\\\\\", "\\\\end{tabular}")
ltx_resize(x)
}
}
