% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{write_tex}
\alias{write_tex}
\title{Write latex code to tex file}
\usage{
write_tex(x, path)
}
\arguments{
\item{x}{object to write to tex}

\item{path}{Character vector with path to file}
}
\value{
\code{write_tex} returns the latex code invisibly (so it can be used
in a pipeline). It also writes x to a tex file.
}
\description{
\code{write_tex} is a function designed to take output from stargazer or
other functions that produce latex code in a character vector and write
it to a tex file.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
